Check for conflicting methods using erasure
Fixed issue making it possible to declare a method with the same erased
signature as an inherited method, yet which did not override that method.

This was fixed by adding an explicit check for inherited methods with
conflicting erased signatures. The fix uses a new attribute
TypeDecl.erasedAnestorMethodsMap.

fixes #103 (bitbucket)
fixes #100 (bitbucket)