Refactor CodeGeneration helper class
Changed the way jump offset errors are handled in CodeGeneration. An exception
is now thrown by CodeGeneration if a jump instruction had a too wide jump
offset that will not fit in two bytes, instead of setting a flag that must be
checked after generating bytecode to see if a jump instruction failed to
generate correctly.

The new thrown exception type is CodeGeneration.JumpOffsetError.