Remove obsolete Java 4 transform
The Java 4 module used side effects to generate methods
for each ClassAccess. This side effect driven transform
was not used in the Java 5+ modules and has been
removed. This may cause problems for Java 4 bytecode
generation, but Java 4 bytecode generation is not tested
for ExtendJ and should be considered obsolete and
unsupported.