Improve enhanced for type checking
Rewrote the enhanced for type checking to use an attribute to lookup the type
argument of the implemented Iterable interface, instead of looking up the
iterable() method and then looking up the next() method to get the element
type.

The new implementation is safer because it does not rely on the ordering of the
iterator() and next() methods in method lookups in case there are multiple
overloaded methods.