Fixed abstract grammar error
There was a missing NTA equation for the NTA BodyDecl list in the PrimitiveType
node. There was also an unused list of BodyDecl nodes in PrimitiveType named
ParsedDeclList.  The unused ParsedDecl:BodyDecl* component of PrimitiveType has
been removed, and the BodyDecl list component has been changed to a non-NTA
list.

Also changed LambdaAnonymousDecl to not use an NTA implements list.