Separate error handling for default constructors
Added separate error messages for default constructor problems. This includes
testing if the default constructor throws a checked exception, and if the
super constructor used in the default constructor is accessible.

Having separate checks for default constructor problems removes the need to
perform full error checking on default constructors, and it means the errors
reported can be more specific.