Improve type hierarchy checking
Added an attribute to test if a type has an enclosing type which is a subtype
of a given type. This is used when checking that a class extending an inner
class is inside a class which is a subtype of the enclosing class of the inner
class it is extending.

Fixes these tests:

* classes/extends_inner_anon_07p