Fix circularity in method reference type analysis
This fixes stack overflow errors in method and constructor reference type
analysis.  The errors occurred because the MethodReference.type() and
ConstructorReference.type() attributes were effectively circular in a when
method or constructor references were used in a type inference context, for
example as argument to a generic method.

fixes #180 (bitbucket)