Improve method type inference for Java 8
Method type inference in Java 8 added inference of the return type, which in a
naive implementation leads to circular type inference when the inferred method
access is inside another inferred method access.

To solve this issue, the type inference for methods has been changed in the
Java 8 extension so that a NTA bound method access is used to perform type
inference.  This effectively pins a method access to a potentially applicable
method so that type inference can be performed on its arguments without
circularly invoking type inference on the outer method access.

This fix is important as many useful methods in the Java 8 standard library
rely the improved method type inference.  For example, the excerpt will now
compile:

	Stream<Integer> stream = ...;
    stream.collect(Collectors.toList());

fixes #182 (bitbucket)