Fix error in ParseName resolution
A rewrite condition has been added to the ParseName rewrite to enforce
left-to-right rewrite ordering in qualified expressions. This fixes a case
where name resolution failed because a right hand side rewrite was triggered
prematurely by an inherited attribute evaluation.

The ParseName was effectively not monotonic, since it could get "stuck" in the
wrong state. This was a very rare occurrence, and only happened in certain cases,
similar to this expression tree:

((String) ((A.B) B).x).length()

The ParseName rewrite of A.B resulted in premature rewriting of x. The
ParseName x was then incorrectly resolved as a package name.

fixes #205 (bitbucket)