Revert "Skip adding redundant equality constraints"
This reverts commit 3ce541ad8f081e4e40f6eecf2c54b2809f3251c5.

In type inference it is possible that a constraint contains the same type
variable on the left and right hand side of the constraint.
For example, T = T.  This can happen when the generic method call is
recursive, calling the same method. In this case, one type variable
belongs to the generic method call, and the other represents a type
from the context of the method call. Even though they are the exact same
type variable, they represent slightly different things and a constraint like
T = T can not be removed because it removes information from the
type inference.

fixes #214