Fix copy-paste error in TypeDecl.typeProblems()
An error check was missing in TypeDecl.typeProblems() in the Java 8 extension
because it had been omitted from the Java 5 implementation. The whole attribute
is re-implemented in the Java 8 extension in order to alter the inherited
Java 4 behavior. This has lead to unintended discrepancies for the Java 8
version.

Also synced error messages to have equal wording in Java 4/5/8.

This fixes regression tests:

* enum/final_07f
* enum/final_09f
* generics/override_02f
* generics/override_10f
* generics/override_12f