Remove unnecessary duplicated file: Constant.jadd
The Java 7 module had a mostly identical copy of java4/frontend/Constant.jadd.
The only real difference was that the Java 7 version had removed the
Constant.buildLiteral() method. This method was not used anywhere so it was
removed from the Java 4 version and then the Java 7 version was deleted.

Also fixed compile error when building only the Java 4 module.