Fix stack map frame generation issue
An incorrect number of assigned local slots was computed when creating a diff
stack frame.  This lead to an incorrect number of append_frame entries.

Also fixed an error in how local variable space was allocated. For wide types,
the second entry was sometimes not set to TOP. Wide types should instead always
be followed by TOP in the local variable map.

fixes #238 (bitbucket)