Fix inner types annotation bytecode parsing issue
Annotations could be dropped when parsing bytecode for inner types.  This
occurred because the modifiers for the inner type were replaced by the
modifiers from the InnerClasses attribute.

This was an order-dependent issue: if the InnerClasses attribute was before the
RuntimeVisibleAnnotations attribute or RuntimeInvisibleAnnotations attribute,
then the annotations would be appended to the updated modifiers.

fixes #256 (bitbucket)