Use Open Source Initiative 3-Clause BSD License
This updates the license text and copyright year in the LICENSE file.  The
license text for the 3-Clause BSD License from The Open Source Initiative is
now used (https://opensource.org/licenses/BSD-3-Clause).

The original LICENSE text was from an unknown source, and did not exactly match
the text from The Open Source Initiative. The license text used in source files
is identical to The OSI version.  The 3-Clause BSD License is also known the
Modified BSD License.